structure the hybrid pkg
